+++ staticpipes/jinja2_environment.py
@@ -12,10 +12,23 @@ class Jinja2Environment:
         self._autoescape = autoescape
         self._filters = filters
 
-    def get(self, source_directory: SourceDirectory) -> jinja2.Environment:
+    def get(
+        self, source_directory: SourceDirectory, secondary_source_directories: dict
+    ) -> jinja2.Environment:
         if not self._jinja2_environment:
+            loaders: list = [jinja2.FileSystemLoader(source_directory.dir)]
+            if secondary_source_directories:
+                loaders.append(
+                    jinja2.PrefixLoader(
+                        {
+                            k: jinja2.FileSystemLoader(ssd.dir)
+                            for k, ssd in secondary_source_directories.items()
+                        },
+                        delimiter=":",
+                    )
+                )
             self._jinja2_environment = jinja2.Environment(
-                loader=jinja2.FileSystemLoader(source_directory.dir),
+                loader=jinja2.ChoiceLoader(loaders),
                 autoescape=self._autoescape,
             )
             for k, v in self._filters.items():