+++ staticpipes/pipes/collection_records_process.py
@@ -74,6 +74,7 @@ class PipeCollectionRecordsProcess(BasePipe):
         for processor in self._processors:
             processor.config = self.config
             processor.source_directory = self.source_directory
+            processor.secondary_source_directories = self.secondary_source_directories
             processor.build_directory = self.build_directory
 
         collection = current_info.get_context("collection")[self._collection_name]